refactor: PlayerBetInfo 클래스명을 BetRecord로 변경 및 BetAmount과 함께 betting 패키지로 이동
